weakdeps for everything, except Json
