+++ Project.toml
@@ -3,20 +3,24 @@ uuid = "db9b398d-9517-45f8-9a95-92af99003e0e"
 version = "4.0.0"
 
 [deps]
-CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-EzXML = "8f5d6c58-4d21-5cfd-889c-e3ad7ee6a615"
-TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 UUIDs = "cf7118a7-6976-5b1a-9a39-7adc72f591a4"
-YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 YYJSON = "dad5a794-77f7-4556-b071-d80a9747599c"
 
 [weakdeps]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+EzXML = "8f5d6c58-4d21-5cfd-889c-e3ad7ee6a615"
 Parquet2 = "98572fba-bba0-415d-956f-fa77e587d26d"
+TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
+YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 
 [extensions]
+SerdeCsvExt = ["CSV"]
 SerdeParquetExt = ["Parquet2", "Tables"]
+SerdeTomlExt = ["TOML"]
+SerdeXmlExt = ["EzXML"]
+SerdeYamlExt = ["YAML"]
 
 [compat]
 CSV = "0.10"
